plans page progress
