+++ packages/files-ui/src/Components/Modules/Settings/ChangePlan/SelectPlan/Products.tsx
@@ -1,11 +1,11 @@
 import React, { useState } from "react"
 import { makeStyles, createStyles } from "@chainsafe/common-theme"
-import { CSFTheme } from "../../../../Themes/types"
+import { CSFTheme } from "../../../../../Themes/types"
 import ProductInfo from "./ProductInfo"
 import { ArrowLeftIcon, Link, RadioInput, Typography } from "@chainsafe/common-components"
 import { t, Trans } from "@lingui/macro"
-import { ROUTE_LINKS } from "../../../FilesRoutes"
-import { useFilesApi } from "../../../../Contexts/FilesApiContext"
+import { ROUTE_LINKS } from "../../../../FilesRoutes"
+import { useFilesApi } from "../../../../../Contexts/FilesApiContext"
 import { Product } from "@chainsafe/files-api-client"
 import { useEffect } from "react"
